refactor: sale timer 로직 utils로 파일 분리
